+++ src/server/components/lists/find/helpers/normaliseServiceType.ts
@@ -1,0 +1,5 @@
+import { kebabCase } from "lodash";
+
+export function normaliseServiceType(serviceType: string) {
+  return kebabCase(serviceType).toLowerCase();
+}
