+++ src/server/components/lists/find/middleware/initFindSession.ts
@@ -1,0 +1,6 @@
+import type { NextFunction, Request, Response } from "express";
+
+export function initFindSession(req: Request, _res: Response, next: NextFunction) {
+  req.session.answers ??= {};
+  next();
+}
