+++ src/server/components/lists/find/middleware/loadCsrfTokenIntoLocals.ts
@@ -1,0 +1,6 @@
+import type { NextFunction, Request, Response } from "express";
+
+export function loadCsrfTokenIntoLocals(req: Request, res: Response, next: NextFunction) {
+  res.locals.csrfToken = req?.csrfToken?.() ?? "";
+  next();
+}
